+++ Documentation/com/llevame_app_project/AsyncLoginTask.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>android.os.AsyncTask&lt;java.lang.String,java.lang.String,java.lang.Boolean&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.llevame_app_project.AsyncLoginTask</li>
+<li>com.llevame_app_project.UserManagement.Registration.Login.AsyncLoginTask</li>
 </ul>
 </li>
 </ul>
