+++ Documentation/com/llevame_app_project/LoginValidator.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.llevame_app_project.LoginValidator</li>
+<li>com.llevame_app_project.UserManagement.Registration.Login.LoginValidator</li>
 </ul>
 </li>
 </ul>
