+++ Documentation/com/llevame_app_project/TravelFragment.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v4.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.llevame_app_project.TravelFragment</li>
+<li>com.llevame_app_project.Activities.TravelFragment</li>
 </ul>
 </li>
 </ul>
