+++ app/src/main/java/com/llevame_app_project/Activities/ProfileActivity.java
@@ -7,7 +7,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.llevame_app_project.Profile;
+import com.llevame_app_project.UserManagement.LoggedUser.Profile;
 import com.llevame_app_project.R;
 
 public class ProfileActivity extends AppCompatActivity {
