+++ app/src/main/java/com/llevame_app_project/Activities/RegistrationActivity.java
@@ -48,8 +48,6 @@ public class RegistrationActivity extends AppCompatActivity {
                 startActivity(intent);
             }
         });
-
-
     }
 
     public void onRadioButtonClicked(View view) {
