+++ app/src/main/java/com/llevame_app_project/UserManagement/Login/LoginValidator.java
@@ -1,4 +1,4 @@
-package com.llevame_app_project;
+package com.llevame_app_project.UserManagement.Login;
 
 public class LoginValidator {
