+++ app/src/main/java/com/llevame_app_project/UserManagement/Registration/PassengerRegistrationThread.java
@@ -1,4 +1,4 @@
-package com.llevame_app_project;
+package com.llevame_app_project.UserManagement.Registration;
 
 import com.llevame_app_project.Data.LoginResponseData;
 import com.llevame_app_project.Data.PassengerData;
