+++ app/src/main/java/com/llevame_app_project/UserManagement/Registration/Registrant.java
@@ -1,4 +1,4 @@
-package com.llevame_app_project;
+package com.llevame_app_project.UserManagement.Registration;
 
 import com.llevame_app_project.Data.CarData;
 import com.llevame_app_project.Data.DriverData;
@@ -6,8 +6,6 @@ import com.llevame_app_project.Data.LoginResponseData;
 import com.llevame_app_project.Data.PassengerData;
 import com.llevame_app_project.Forms.FirstRegistrationForm;
 import com.llevame_app_project.Forms.SecondRegistrationForm;
-
-import java.sql.Driver;
 
 import retrofit2.Response;
 
@@ -82,6 +80,4 @@ public class Registrant {
         driverData.setCar(carData);
         return driverData;
     }
-
-
 }
