+++ app/src/test/java/com/llevame_app_project/LoginValidatorTest.java
@@ -1,17 +1,20 @@
 package com.llevame_app_project;
+
+import com.llevame_app_project.UserManagement.Login.LoginValidator;
+
 import org.junit.Test;
 
 //Local unit test, which will execute on the development machine (host).
 
 public class LoginValidatorTest {
     @Test
-    public void anEmailIsNotvalidWithoutAtSymbol(){
+    public void anEmailIsNotValidWithoutAtSymbol(){
         LoginValidator loginValidator = new LoginValidator();
         assert(loginValidator.isEmailValid("a@b.com"));
     }
 
     @Test
-    public void aPasswordlIsNotvalidIfItIsShorterThan4Characters(){
+    public void aPasswordIsNotValidIfItIsShorterThan4Characters(){
         LoginValidator loginValidator = new LoginValidator();
         assert(loginValidator.isPasswordValid("abc"));
     }
